+++ cognite/neat/_issues/_base.py
@@ -1,15 +1,12 @@
 import inspect
 import sys
 import warnings
-from abc import ABC
-from collections.abc import Collection, Hashable, Iterable, Iterator, Sequence
-from contextlib import contextmanager
+from collections.abc import Collection, Hashable, Iterable, Sequence
 from dataclasses import dataclass, fields
 from functools import total_ordering
 from pathlib import Path
 from types import UnionType
 from typing import Any, ClassVar, Literal, TypeAlias, TypeVar, get_args, get_origin
-from warnings import WarningMessage
 
 import pandas as pd
 from cognite.client.data_classes.data_modeling import (
@@ -18,10 +15,7 @@ from cognite.client.data_classes.data_modeling import (
     PropertyId,
     ViewId,
 )
-from pydantic import ValidationError
-from pydantic_core import ErrorDetails
 
-from cognite.neat._utils.spreadsheet import SpreadsheetRead
 from cognite.neat._utils.text import humanize_collection, to_camel_case, to_snake_case
 
 if sys.version_info < (3, 11):
@@ -32,11 +26,10 @@ else:
 
 
 __all__ = [
-    "DefaultWarning",
+    "IssueList",
     "MultiValueError",
     "NeatError",
     "NeatIssue",
-    "NeatIssueList",
     "NeatWarning",
 ]
 
@@ -210,212 +203,42 @@ class NeatIssue:
             return NotImplemented
         return (type(self).__name__, self.as_message()) == (type(other).__name__, other.as_message())
 
+    def __str__(self) -> str:
+        return self.as_message()
+
 
 @dataclass(unsafe_hash=True)
 class NeatError(NeatIssue, Exception):
     """This is the base class for all exceptions (errors) used in Neat."""
 
-    @classmethod
-    def from_errors(cls, errors: "list[ErrorDetails | NeatError]", **kwargs) -> "list[NeatError]":
-        """Convert a list of pydantic errors to a list of Error instances.
-
-        This is intended to be overridden in subclasses to handle specific error types.
-        """
-        all_errors: list[NeatError] = []
-        read_info_by_sheet = kwargs.get("read_info_by_sheet")
-
-        for error in errors:
-            if (
-                isinstance(error, dict)
-                and error["type"] == "is_instance_of"
-                and error["loc"][1] == "is-instance[SheetList]"
-            ):
-                # Skip the error for SheetList, as it is not relevant for the user. This is an
-                # internal class used to have helper methods for a lists as .to_pandas()
-                continue
-
-            neat_error: NeatError | None = None
-            if isinstance(error, dict) and isinstance(ctx := error.get("ctx"), dict) and "error" in ctx:
-                neat_error = ctx["error"]
-            elif isinstance(error, NeatError | MultiValueError):
-                neat_error = error
-
-            loc = error["loc"] if isinstance(error, dict) else tuple()
-            if isinstance(neat_error, MultiValueError):
-                all_errors.extend([cls._adjust_error(e, loc, read_info_by_sheet) for e in neat_error.errors])
-            elif isinstance(neat_error, NeatError):
-                all_errors.append(cls._adjust_error(neat_error, loc, read_info_by_sheet))
-            elif isinstance(error, dict) and len(loc) >= 4 and read_info_by_sheet:
-                all_errors.append(RowError.from_pydantic_error(error, read_info_by_sheet))
-            elif isinstance(error, dict):
-                all_errors.append(DefaultPydanticError.from_pydantic_error(error))
-            else:
-                # This is unreachable. However, in case it turns out to be reachable, we want to know about it.
-                raise ValueError(f"Unsupported error type: {error}")
-        return all_errors
-
-    @classmethod
-    def _adjust_error(
-        cls, error: "NeatError", loc: tuple[str | int, ...], read_info_by_sheet: dict[str, SpreadsheetRead] | None
-    ) -> "NeatError":
-        from .errors._wrapper import MetadataValueError
-
-        if read_info_by_sheet:
-            cls._adjust_row_numbers(error, read_info_by_sheet)
-        if len(loc) == 2 and isinstance(loc[0], str) and loc[0].casefold() == "metadata":
-            return MetadataValueError(field_name=str(loc[1]), error=error)
-        return error
-
-    @staticmethod
-    def _adjust_row_numbers(caught_error: "NeatError", read_info_by_sheet: dict[str, SpreadsheetRead]) -> None:
-        from cognite.neat._issues.errors._properties import PropertyDefinitionDuplicatedError
-        from cognite.neat._issues.errors._resources import ResourceNotDefinedError
-
-        reader = read_info_by_sheet.get("Properties", SpreadsheetRead())
-
-        if isinstance(caught_error, PropertyDefinitionDuplicatedError) and caught_error.location_name == "rows":
-            adjusted_row_number = (
-                tuple(
-                    reader.adjusted_row_number(row_no) if isinstance(row_no, int) else row_no
-                    for row_no in caught_error.locations or []
-                )
-                or None
-            )
-            # The error is frozen, so we have to use __setattr__ to change the row number
-            object.__setattr__(caught_error, "locations", adjusted_row_number)
-        elif isinstance(caught_error, RowError):
-            # Adjusting the row number to the actual row number in the spreadsheet
-            new_row = reader.adjusted_row_number(caught_error.row)
-            # The error is frozen, so we have to use __setattr__ to change the row number
-            object.__setattr__(caught_error, "row", new_row)
-        elif isinstance(caught_error, ResourceNotDefinedError):
-            if isinstance(caught_error.row_number, int) and caught_error.sheet_name == "Properties":
-                new_row = reader.adjusted_row_number(caught_error.row_number)
-                object.__setattr__(caught_error, "row_number", new_row)
-
-
-@dataclass(unsafe_hash=True)
-class DefaultPydanticError(NeatError, ValueError):
-    """{type}: {msg} [loc={loc}]"""
-
-    type: str
-    loc: tuple[int | str, ...]
-    msg: str
-
-    @classmethod
-    def from_pydantic_error(cls, error: ErrorDetails) -> "NeatError":
-        loc = error["loc"]
-        if len(loc) >= 2 and isinstance(loc[0], str) and loc[0].casefold() == "metadata":
-            from .errors._general import NeatValueError
-            from .errors._wrapper import MetadataValueError
-
-            return MetadataValueError(
-                field_name=str(loc[1]), error=NeatValueError(f"{error['msg']} got '{error['input']}'")
-            )
-
-        return cls(
-            type=error["type"],
-            loc=error["loc"],
-            msg=error["msg"],
-        )
-
-    def as_message(self, include_type: bool = True) -> str:
-        if self.loc and len(self.loc) == 1:
-            return f"{self.loc[0]} sheet: {self.msg}"
-        elif self.loc and len(self.loc) == 2:
-            return f"{self.loc[0]} sheet field/column <{self.loc[1]}>: {self.msg}"
-        else:
-            return self.msg
-
-
-@dataclass(unsafe_hash=True)
-class RowError(NeatError, ValueError):
-    """In {sheet_name}, row={row}, column={column}: {msg}. [type={type}, input_value={input}]"""
-
-    extra = "For further information visit {url}"
-
-    sheet_name: str
-    column: str
-    row: int
-    type: str
-    msg: str
-    input: Any
-    url: str | None = None
-
-    @classmethod
-    def from_pydantic_error(
-        cls,
-        error: ErrorDetails,
-        read_info_by_sheet: dict[str, SpreadsheetRead] | None = None,
-    ) -> Self:
-        sheet_name, _, row, column, *__ = error["loc"]
-        reader = (read_info_by_sheet or {}).get(str(sheet_name), SpreadsheetRead())
-        return cls(
-            sheet_name=str(sheet_name),
-            column=str(column),
-            row=reader.adjusted_row_number(int(row)),
-            type=error["type"],
-            msg=error["msg"],
-            input=error.get("input"),
-            url=str(url) if (url := error.get("url")) else None,
-        )
-
-    def as_message(self, include_type: bool = True) -> str:
-        input_str = str(self.input) if self.input is not None else ""
-        input_str = input_str[:50] + "..." if len(input_str) > 50 else input_str
-        output = (
-            f"In {self.sheet_name}, row={self.row}, column={self.column}: {self.msg}. "
-            f"[type={self.type}, input_value={input_str}]"
-        )
-        if self.url:
-            output += f" For further information visit {self.url}"
-        return output
+    ...
 
 
 @dataclass(unsafe_hash=True)
 class NeatWarning(NeatIssue, UserWarning):
     """This is the base class for all warnings used in Neat."""
 
-    @classmethod
-    def from_warning(cls, warning: WarningMessage) -> "NeatWarning":
-        """Create a NeatWarning from a WarningMessage."""
-        return DefaultWarning.from_warning_message(warning)
+    ...
 
 
-@dataclass(unsafe_hash=True)
-class DefaultWarning(NeatWarning):
-    """{category}: {warning}"""
+class MultiValueError(ValueError):
+    """This is a container for multiple errors.
 
-    extra = "Source: {source}"
+    It is used in the pydantic field_validator/model_validator to collect multiple errors, which
+    can then be caught in a try-except block and returned as an IssueList.
 
-    warning: str
-    category: str
-    source: str | None = None
+    """
 
-    @classmethod
-    def from_warning_message(cls, warning: WarningMessage) -> NeatWarning:
-        if isinstance(warning.message, NeatWarning):
-            return warning.message
-
-        return cls(
-            warning=str(warning.message),
-            category=warning.category.__name__,
-            source=warning.source,
-        )
-
-    def as_message(self, include_type: bool = True) -> str:
-        return str(self.warning)
+    def __init__(self, errors: Sequence[NeatIssue]):
+        self.errors = IssueList(errors)
 
 
-T_NeatIssue = TypeVar("T_NeatIssue", bound=NeatIssue)
-
-
-class NeatIssueList(list, Sequence[T_NeatIssue], ABC):
+class IssueList(list, Sequence[NeatIssue]):
     """This is a generic list of NeatIssues."""
 
     def __init__(
         self,
-        issues: Sequence[T_NeatIssue] | None = None,
+        issues: Sequence[NeatIssue] | None = None,
         title: str | None = None,
         action: str | None = None,
         hint: str | None = None,
@@ -462,35 +285,16 @@ class NeatIssueList(list, Sequence[T_NeatIssue], ABC):
 
     def trigger_warnings(self) -> None:
         """Trigger all warnings in this list."""
-        for warning in [issue for issue in self if isinstance(issue, NeatWarning)]:
+        for warning in self.warnings:
             warnings.warn(warning, stacklevel=2)
 
     def to_pandas(self) -> pd.DataFrame:
         """Return a pandas DataFrame representation of this list."""
         return pd.DataFrame([issue.dump() for issue in self])
 
-    def _repr_html_(self) -> str | None:
-        return self.to_pandas()._repr_html_()  # type: ignore[operator]
-
-    def as_exception(self) -> "MultiValueError":
+    def as_exception(self) -> MultiValueError:
         """Return a MultiValueError with all the errors in this list."""
         return MultiValueError(self.errors)
-
-
-class MultiValueError(ValueError):
-    """This is a container for multiple errors.
-
-    It is used in the pydantic field_validator/model_validator to collect multiple errors, which
-    can then be caught in a try-except block and returned as an IssueList.
-
-    """
-
-    def __init__(self, errors: Sequence[NeatIssue]):
-        self.errors = list(errors)
-
-
-class IssueList(NeatIssueList[NeatIssue]):
-    """This is a list of NeatIssues."""
 
     def _repr_html_(self) -> str | None:
         if self.action and not self:
@@ -530,38 +334,3 @@ def _get_subclasses(cls_: type[T_Cls], include_base: bool = False) -> Iterable[t
     for s in cls_.__subclasses__():
         yield s
         yield from _get_subclasses(s, False)
-
-
-@contextmanager
-def catch_warnings() -> Iterator[IssueList]:
-    """Catch warnings and append them to the issues list."""
-    issues = IssueList()
-    with warnings.catch_warnings(record=True) as warning_logger:
-        warnings.simplefilter("always")
-        try:
-            yield issues
-        finally:
-            if warning_logger:
-                issues.extend([NeatWarning.from_warning(warning) for warning in warning_logger])  # type: ignore[misc]
-
-
-@contextmanager
-def catch_issues(error_args: dict[str, Any] | None = None) -> Iterator[IssueList]:
-    """This is an internal help function to handle issues and warnings.
-
-    Args:
-        error_args: Additional arguments to pass to the error class. The only use case as of (2025-01-03) is to pass
-            the read_info_by_sheet to the error class such that the row numbers can be adjusted to match the source
-            spreadsheet.
-
-    Returns:
-        IssueList: The list of issues.
-
-    """
-    with catch_warnings() as issues:
-        try:
-            yield issues
-        except ValidationError as e:
-            issues.extend(NeatError.from_errors(e.errors(), **(error_args or {})))  # type: ignore[arg-type]
-        except (NeatError, MultiValueError) as e:
-            issues.extend(NeatError.from_errors([e], **(error_args or {})))  # type: ignore[arg-type, list-item]