+++ cognite/neat/_issues/errors/__init__.py
@@ -1,4 +1,4 @@
-from cognite.neat._issues._base import DefaultPydanticError, NeatError, RowError, _get_subclasses
+from cognite.neat._issues._base import NeatError, _get_subclasses
 
 from ._external import (
     AuthorizationError,
@@ -31,12 +31,23 @@ from ._resources import (
     ResourceNotFoundError,
     ResourceRetrievalError,
 )
-from ._wrapper import MetadataValueError
+from ._wrapper import (
+    ClassValueError,
+    ContainerValueError,
+    EnumValueError,
+    MetadataValueError,
+    NodeValueError,
+    PropertyValueError,
+    SpreadsheetError,
+    ViewValueError,
+)
 
 __all__ = [
     "AuthorizationError",
     "CDFMissingClientError",
-    "DefaultPydanticError",
+    "ClassValueError",
+    "ContainerValueError",
+    "EnumValueError",
     "FileMissingRequiredFieldError",
     "FileNotAFileError",
     "FileNotFoundNeatError",
@@ -48,12 +59,14 @@ __all__ = [
     "NeatTypeError",
     "NeatValueError",
     "NeatYamlError",
+    "NodeValueError",
     "OxigraphStorageLockedError",
     "PropertyDefinitionDuplicatedError",
     "PropertyDefinitionError",
     "PropertyMappingDuplicatedError",
     "PropertyNotFoundError",
     "PropertyTypeNotSupportedError",
+    "PropertyValueError",
     "RegexViolationError",
     "ResourceChangedError",
     "ResourceConversionError",
@@ -65,7 +78,8 @@ __all__ = [
     "ResourceNotFoundError",
     "ResourceRetrievalError",
     "ReversedConnectionNotFeasibleError",
-    "RowError",
+    "SpreadsheetError",
+    "ViewValueError",
 ]
 
 _NEAT_ERRORS_BY_NAME = {error.__name__: error for error in _get_subclasses(NeatError, include_base=True)}