+++ tests/tests_unit/rules/test_importers/test_excel_importer.py
@@ -2,14 +2,15 @@ from pathlib import Path
 
 import pytest
 from cognite.client.data_classes.data_modeling import ContainerId, ViewId
-from pydantic.version import VERSION
 
 from cognite.neat._issues import IssueList, catch_issues
 from cognite.neat._issues.errors import (
     CDFMissingClientError,
     FileNotFoundNeatError,
+    MetadataValueError,
+    NeatValueError,
     PropertyDefinitionDuplicatedError,
-    RowError,
+    PropertyValueError,
 )
 from cognite.neat._issues.warnings import (
     NotSupportedHasDataFilterLimitWarning,
@@ -28,21 +29,29 @@ def invalid_rules_filepaths():
         IssueList([FileNotFoundNeatError(DOC_RULES / "not-existing.xlsx")]),
         id="Not existing file",
     )
-    major, minor, *_ = VERSION.split(".")
+
+    yield pytest.param(
+        EXCEL_IMPORTER_DATA / "invalid_metadata.xlsx",
+        IssueList(
+            [
+                MetadataValueError(
+                    field_name="space",
+                    error=NeatValueError("value is missing."),
+                ),
+            ]
+        ),
+        id="Missing space in Metadata sheet.",
+    )
 
     yield pytest.param(
         EXCEL_IMPORTER_DATA / "invalid_property_dms_rules.xlsx",
         IssueList(
             [
-                RowError(
-                    sheet_name="Properties",
-                    column="Is List",
+                PropertyValueError(
                     row=5,
-                    type="bool_parsing",
-                    msg="Input should be a valid boolean, unable to interpret input",
-                    input="Apple",
-                    url=f"https://errors.pydantic.dev/{major}.{minor}/v/bool_parsing",
-                )
+                    column="Is List",
+                    error=NeatValueError("Expected a bool type, got 'Apple'"),
+                ),
             ]
         ),
         id="Invalid property specification",